polys: Moved matrix_fglm() to its own module
